Add error handling framework, security improvements, and development infrastructure
- Add comprehensive error handling system with ErrorHandler and ErrorLogger utilities
- Implement security improvements and safe process execution utilities
- Add pre-commit hooks configuration and secrets baseline scanning
- Enhance development environment with Docker ARM64 support and setup scripts
- Improve code quality with updated clang-format and clang-tidy configurations
- Add MVP3 planning documentation and development requirements
- Enhance testing infrastructure with coverage and sanitizer builds
- Update build system with improved CMake and Ninja configuration

+++ src/cli/commands/doctor_command.cpp
@@ -457,7 +457,7 @@ Result<DoctorCommand::HealthCheck> DoctorCommand::checkExternalTools() {
     };
     
     int available = 0;
-    int total = tools.size();
+    int total = static_cast<int>(tools.size());
     std::string missing_tools;
     
     for (const auto& [tool, exists] : tools) {
@@ -633,12 +633,12 @@ Result<DoctorCommand::HealthCheck> DoctorCommand::checkNotesIntegrity() {
     } else {
       auto notes = notes_result.value();
       int corrupted = 0;
-      int total = notes.size();
+      int total = static_cast<int>(notes.size());
       
       // Sample a few notes to check for corruption
       int to_check = std::min(10, total);
       for (int i = 0; i < to_check && i < total; ++i) {
-        auto note_result = note_store.load(notes[i]);
+        auto note_result = note_store.load(notes[static_cast<size_t>(i)]);
         if (!note_result.has_value()) {
           corrupted++;
         }
